zeldadungeon
Each room should contain a puzzle or combat encounter (or something else that makes it interesting)
The player should start without their weapon, and be required to collect their weapon to complete the dungeon. This weapon is the "Key Item" for the dungeon (represented in the graphs with a Bow and Arrow). After getting the weapon, the player can hit enemies and break boxes.
Just to make sure it's clear: your Whitebox should have gameplay elements including keys, buttons, moving platforms, enemies, etc.